+++ pkg/keyvisual/decorator/decorator_test.go
