+++ pkg/keyvisual/decorator/utils.go
