+++ pkg/keyvisual/info/interface.go
