+++ pkg/keyvisual/input/file.go
@@ -20,7 +20,7 @@ import (
 	"github.com/pingcap/log"
 	"go.uber.org/zap"
 
-	"github.com/pingcap-incubator/tidb-dashboard/pkg/apiserver/keyvisual/storage"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/storage"
 )
 
 type fileInput struct {
