+++ pkg/keyvisual/input/periodic.go
@@ -20,8 +20,8 @@ import (
 	"github.com/pingcap/log"
 	"go.uber.org/zap"
 
-	"github.com/pingcap-incubator/tidb-dashboard/pkg/apiserver/keyvisual/info"
-	"github.com/pingcap-incubator/tidb-dashboard/pkg/apiserver/keyvisual/storage"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/info"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/storage"
 )
 
 type periodicInput struct {
