+++ pkg/keyvisual/matrix/average_test.go
