+++ pkg/keyvisual/matrix/distance.go
@@ -19,8 +19,8 @@ import (
 	"sort"
 	"sync"
 
-	"github.com/pingcap-incubator/tidb-dashboard/pkg/apiserver/keyvisual/decorator"
 	"github.com/pingcap-incubator/tidb-dashboard/pkg/config"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/decorator"
 )
 
 // TODO:
