+++ pkg/keyvisual/matrix/util.go
