+++ pkg/keyvisual/storage/stat_test.go
